+++ frontend/src/lib/SVGIcons/ReorderIcon.svelte
@@ -1,15 +1,16 @@
 <script lang="ts">
 </script>
 
-<svg viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+<svg viewBox="0 0 23 24" fill="none" xmlns="http://www.w3.org/2000/svg">
   <path d="M17 15L12 20L7 15" stroke="#33363F" stroke-width="2" />
   <path d="M17 9L12 4L7 9" stroke="#33363F" stroke-width="2" />
 </svg>
 
 <style>
   svg {
-    width: 1rem;
-    height: 1rem;
+    width: 1.1rem;
+    height: 1.1rem;
+    margin-top: 0.2rem;
     stroke: var(--text-muted);
     transition: stroke 0.4s;
   }